+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for AWS API Gateway Lambda Proxy Integration module
 
+## [1.0.3] - 2018-01-19
+* Added CloudWatch Logs link generation for error responses. By default the
+error response now has a "log" attribute that contains a direct link to the
+logs of the failing request.
+
 ## [1.0.2] - 2018-01-18
 * Fixed package dependencies and setup Travis build.